Getting things ready for action builder More SIMD goodness
Former-commit-id: e6e9df65ffd48efa2c3481f1c3094b787a533fb3

+++ Exporters/3ds Max/Max2Babylon/Exporter/BabylonExporter.Animation.cs
@@ -154,11 +154,11 @@ namespace Max2Babylon
 
             var babylonAnimation = new BabylonAnimation
             {
-                dataType = dataType,
+                dataType = (int)dataType,
                 name = property + " animation",
                 keys = keys.ToArray(),
                 framePerSecond = Loader.Global.FrameRate,
-                loopBehavior = loopBehavior,
+                loopBehavior = (int)loopBehavior,
                 property = property
             };
 
@@ -283,11 +283,11 @@ namespace Max2Babylon
 
                     var babylonAnimation = new BabylonAnimation
                     {
-                        dataType = dataType,
+                        dataType = (int)dataType,
                         name = property + " animation",
                         keys = keys.ToArray(),
                         framePerSecond = Loader.Global.FrameRate,
-                        loopBehavior = BabylonAnimation.LoopBehavior.Cycle,
+                        loopBehavior = (int)BabylonAnimation.LoopBehavior.Cycle,
                         property = property
                     };
